prov/psm: Avoid a long PSM delay when the device is absent
The PSM function that is used to check the number of available device
units for PSM may wait up to 15 seconds for the device to come up.
This introduces unnecessary delay for setups that have the PSM library
installed w/o the actual hardware. This patch provides a workaround.

Port of commit 31384811a549 ("prov/psm2: Avoid a long PSM2 delay when
the device is absent") and commit 5360e8c307ce ("prov/psm2: Improve the
method of checking device existence") from PSM2 to PSM.

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>